# Speed up rendering by determining gray-out colors upfront