--- a/gaphor/diagram/propertypages.glade
+++ b/gaphor/diagram/propertypages.glade
@@ -585,6 +585,7 @@ Use -/= to move items up or down.</property>
               <object class="GtkTextView" id="comment">
                 <property name="visible">True</property>
                 <property name="can_focus">True</property>
+                <property name="wrap_mode">word</property>
               </object>
             </child>
           </object>